Added aliases to export

+++ scripts/convert.php
@@ -13,11 +13,23 @@ if (!$version) {
 }
 
 
-
+if (!is_file($cwd.'/../lib/meta/json_'.$version.'.txt')) {
+    echo('Please run export_modules.js first!');
+    exit;
+}
 $config = json_decode(file_get_contents($cwd.'/../lib/meta/json_'.$version.'.txt'), true);
 
-$out = '<?php $GLOBALS[\'yui_current\'] = ' . var_export($config, true) . '; ?>';
+if (is_file($cwd.'/../lib/meta/json_aliases_'.$version.'.txt')) {
+    $aliases = json_decode(file_get_contents($cwd.'/../lib/meta/json_aliases_'.$version.'.txt'), true);
+}
+
+
+$out = '<?php $GLOBALS[\'yui_current\'] = ' . var_export($config, true) . ';'."\n\n";
+if ($aliases) {
+    $out .= '$GLOBALS[\'yui_current_aliases\'] = ' . var_export($aliases, true) . ';'."\n";
+}
+$out .= ' ?>';
 
 file_put_contents($cwd.'/../lib/meta/config_'.$version.'.php', $out);
-echo('Finished converting JSON to PHP\n');
+echo("Finished converting JSON to PHP\n");
 ?>
